feat(MovieApi): Add a generic to make requests.
